requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
spacy==3.7.2
pdfminer.six==20221105
python-docx==1.1.0
pydantic==2.5.0
pytest==7.4.3
loguru==0.7.2
